TRCLI-230: Added quality rating support via --result-fields option
